Risolti alcuni problemi legati a getter e setter cell.

+++ MeanFormula.cpp
@@ -23,7 +23,7 @@ void MeanFormula::calc() {
     float sum = 0;
     int count = 0;
     for (auto itr = cells.begin(); itr != cells.end(); ++itr) {
-        sum += (*itr)->getValue();
+        sum += (*itr)->getValue().toFloat();
         count++;
     }
     mean = (sum / count);
